pyproject.toml
[project]
name = "finalllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camelot-py[cv]>=0.11.0",
    "chromadb>=0.4.22",
    "langchain>=0.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-text-splitters>=0.0.1",
    "langgraph>=0.0.20",
    "numpy>=1.24.0",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "pdf2image>=1.16.3",
    "pdfplumber>=0.10.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pymupdf>=1.23.0",
    "pypdf>=3.17.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
    "tabula-py>=2.8.0",
    "tqdm>=4.66.0",
    "typing-extensions>=4.5.0",
    "unstructured>=0.10.0",
]
